Update documentation for github updates
