+++ docs/components.md
@@ -81,7 +81,7 @@ Three slots are available. Besides the main slot for the content there is a name
 
 The component exports two properties to determine its behavior. `collapsible` is a boolean describing wether the component should behave like a collapsible in the first place. `open` is a boolean as well and if set true shows the full content of the body on mount.
 
-## Spinner
+## Spinner (OBSOLETE)
 
 A small component showing an animated spinner which can be used while waiting for data.
 
@@ -109,3 +109,11 @@ and call one of the 4 toast methods:
 | `notification.success(msg:string, timeout:number)` | :octicons-check-circle-16: Shows as success message |
 
 Each method takes an `msg`-string as an argument, which will be shown as the message body. It accepts HTML yo enrich your toasts, if you should desire to do so. The `timeout` argument specifies how many milliseconds the toast notification shall be shown to the user.
+
+## Github Update Dialog
+
+This is a modal showing the update progress, possible error messages and makes a full page refresh 5 seconds after the OTA was successful.
+
+## Update Indicator
+
+The update indicator is a small widget shown in the upper right corner of the status bar. It indicates the availability of a newer stable firmware release then the current one. Upon pressing the icon it will automatically update the firmware to the latest stable release. By default this works through the Github Latest Release API. This must be customized should you use a different update server. Have a look at the [source file](https://github.com/theelims/ESP32-sveltekit/blob/main/interface/src/lib/components/GithubUpdateDialog.svelte) to see what portions to update.
