+++ docs/restfulapi.md
@@ -25,8 +25,9 @@ The back end exposes a number of API endpoints which are referenced in the table
 | POST   | /rest/restart             | `IS_ADMIN`         | none                                                                                                                                                                                                                               | Restart the ESP32                                                              |
 | POST   | /rest/factoryReset        | `IS_ADMIN`         | none                                                                                                                                                                                                                               | Reset the ESP32 and all settings to their default values                       |
 | POST   | /rest/uploadFirmware      | `IS_ADMIN`         | none                                                                                                                                                                                                                               | File upload of firmware.bin                                                    |
-| POST   | /rest/signIn              | `NONE_REQUIRED`    | `{"password": "admin","username": "admin"}  `                                                                                                                                                                                      | Signs a user in and returns access token                                       |
+| POST   | /rest/signIn              | `NONE_REQUIRED`    | `{"password": "admin","username": "admin"}`                                                                                                                                                                                        | Signs a user in and returns access token                                       |
 | GET    | /rest/securitySettings    | `IS_ADMIN`         | none                                                                                                                                                                                                                               | retrieves all user information and roles                                       |
 | POST   | /rest/securitySettings    | `IS_ADMIN`         | `{"jwt_secret": "734cb5bb-5597b722", "users": [{"username": "admin", "password": "admin", "admin": true}, {"username": "guest", "password": "guest", "admin": false, }]`                                                           | retrieves all user information and roles                                       |
 | GET    | /rest/verifyAuthorization | `NONE_REQUIRED`    | none                                                                                                                                                                                                                               | Verifies the content of the auth bearer token                                  |
 | POST   | /rest/sleep               | `IS_AUTHENTICATED` | none                                                                                                                                                                                                                               | Puts the device in deep sleep mode                                             |
+| POST   | /rest/downloadUpdate      | `IS_ADMIN`         | `{"download_url": "https://github.com/theelims/ESP32-sveltekit/releases/download/v0.1.0/firmware_esp32s3.bin"}`                                                                                                                    | Download link for OTA                                                          |
