+++ features.ini
@@ -5,7 +5,7 @@ build_flags =
   -D FT_NTP=1
   -D FT_OTA=1
   -D FT_UPLOAD_FIRMWARE=1
-  -D FT_DOWNLOAD_FIRMWARE=1
+  -D FT_DOWNLOAD_FIRMWARE=1 ;requires FT_NTP=1
   -D FT_SLEEP=1
   -D FT_BATTERY=1
