+++ interface/src/routes/+layout.ts
@@ -9,6 +9,7 @@ export const load = (async () => {
 	const item = await result.json();
 	return {
 		features: item,
-		title: 'ESP32-SvelteKit'
+		title: 'ESP32-SvelteKit',
+		github: 'theelims/ESP32-sveltekit'
 	};
 }) satisfies LayoutLoad;
