+++ platformio.ini
@@ -26,9 +26,9 @@ build_flags =
     -D SSE_MAX_QUEUED_MESSAGES=64
     -D WS_MAX_QUEUED_MESSAGES=64
 	; Uncomment to receive log messages from the ESP Arduino Core
-	-DCORE_DEBUG_LEVEL=5
+	-D CORE_DEBUG_LEVEL=5
     ; Move all networking stuff to the protocol core 0 and leave business logic on application core 1
-    -DCONFIG_ASYNC_TCP_RUNNING_CORE=0
+    -D CONFIG_ASYNC_TCP_RUNNING_CORE=0
 	-D NO_GLOBAL_ARDUINOOTA
     ; Uncomment PROGMEM_WWW to enable the storage of the WWW data in PROGMEM
     -D PROGMEM_WWW
@@ -38,7 +38,7 @@ build_flags =
 lib_compat_mode = strict
 
 ; Uncomment to include the Github Root CA SSL Certificate to pull OTA from the github release page
-; Needs -D FT_DOWNLOAD_FIRMWARE=1 
+; Needs -D FT_DOWNLOAD_FIRMWARE=1 and -D FT_NTP=1
 board_build.embed_txtfiles =
   src/certs/github_root_ca.crt
